Variable to skip container interface wiring script
This script is used to allow adding new interfaces without a container
reboot. But when upgrading to Rocky it results in loss of network
connectivity in the container because it takes down all interfaces and
tries to configure them with systemd-networkd. But systemd-networkd is
not yet configured at this point in the upgrade.